requirements.txt
streamlit>=1.28.0
numpy>=1.21.0
python-dotenv>=0.19.0
openai>=1.0.0
streamlit-audiorec>=0.1.3